+++ src/tidytcells/__init__.py
@@ -1,7 +1,3 @@
 """
 Package root.
 """
-
-
-# Directly import submodules
-from . import aa, junction, mhc, tcr
