+++ src/tidytcells/_utils/misc.py
@@ -1,0 +1,2 @@
+def lowercase_and_remove_whitespace(string: str) -> str:
+    return "".join(string.split()).lower()
