+++ src/tidytcells/aa/__init__.py
@@ -3,11 +3,4 @@ Functions to clean and standardize amino acid sequence data.
 """
 
 
-from ._mod import standardize
-
-
-def standardise(*args, **kwargs):
-    """
-    Alias for :py:func:`tidytcells.aa.standardize`.
-    """
-    return standardize(*args, **kwargs)
+from ._standardize import standardize, standardise
