+++ src/tidytcells/junction/__init__.py
@@ -3,11 +3,4 @@ Functions to clean and standardize junction (CDR3) data.
 """
 
 
-from ._mod import standardize
-
-
-def standardise(*args, **kwargs):
-    """
-    Alias for :py:func:`tidytcells.junction.standardize`.
-    """
-    return standardize(*args, **kwargs)
+from ._standardize import standardize, standardise
