+++ src/tidytcells/mhc/_standardize.py
@@ -1,0 +1,143 @@
+from typing import Dict, Optional, Type
+
+from tidytcells import _utils
+from tidytcells._utils import Parameter, warnings
+from tidytcells._utils.gene_standardizers import (
+    GeneStandardizer,
+    HLAStandardizer,
+    MusMusculusMHCStandardizer,
+)
+
+
+STANDARDIZERS: Dict[str, Type[GeneStandardizer]] = {
+    "homosapiens": HLAStandardizer,
+    "musmusculus": MusMusculusMHCStandardizer,
+}
+
+
+def standardize(
+    gene: Optional[str] = None,
+    species: str = "homosapiens",
+    precision: str = "allele",
+    on_fail: str = "reject",
+    suppress_warnings: bool = False,
+    gene_name: Optional[str] = None,
+) -> tuple:
+    """
+    Attempt to standardize an MHC gene name to be IMGT-compliant.
+
+    .. topic:: Supported species
+
+        - ``'homosapiens'``
+        - ``'musmusculus'``
+
+    .. note::
+        This function will only verify the validity of an MHC gene/allele up to the level of the protein.
+        Any further precise allele designations will not be verified, apart from the requirement that the format (colon-separated numbers) look valid.
+        The reasons for this is firstly because new alleles at that level are added to the IMGT list quite often and so accurate verification is difficult, secondly because people rarely need verification to such a precise level, and finally because such verification costs more computational effort with diminishing returns.
+
+    :param gene:
+        Potentially non-standardized MHC gene name.
+    :type gene:
+        str
+    :param species:
+        Species to which the MHC gene belongs (see above for supported species).
+        Defaults to ``'homosapiens'``.
+    :type species:
+        str
+    :param precision:
+        The maximum level of precision to standardize to.
+        ``'allele'`` standardizes to the maximum precision possible.
+        ``'protein'`` keeps allele designators up to the level of the protein.
+        ``'gene'`` standardizes only to the level of the gene.
+        Defaults to ``'allele'``.
+    :type precision:
+        str
+    :param on_fail:
+        Behaviour when standardization fails.
+        If set to ``"reject"``, returns ``None`` on failure.
+        If set to ``"keep"``, returns the original input.
+        Defaults to ``"reject"``.
+    :type on_fail:
+        str
+    :param suppress_warnings:
+        Disable warnings that are usually emitted when standardisation fails.
+        Defaults to ``False``.
+    :type suppress_warnings:
+        bool
+
+    :param gene_name:
+        Alias for the parameter ``gene``.
+
+        .. caution:: This will be deprecated soon in favour of ``gene``.
+    :type gene_name:
+        str
+
+    :return:
+        If the specified ``species`` is supported, and ``gene`` could be standardized, then return the standardized gene name.
+        If ``species`` is unsupported, then the function does not attempt to standardize, and returns the unaltered ``gene`` string.
+        Else follows the behvaiour as set by ``on_fail``.
+    :rtype:
+        Union[str, None]
+
+    .. topic:: Example usage
+
+        Input strings will intelligently be corrected to IMGT-compliant gene symbols.
+
+        >>> tt.mhc.standardize("A1")
+        'HLA-A*01'
+
+        The ``precision`` setting can truncate unnecessary information.
+
+        >>> tt.mhc.standardize("HLA-A*01", precision="gene")
+        'HLA-A'
+
+        *Mus musculus* is a supported species.
+
+        >>> tt.mhc.standardize("CRW2", species="musmusculus")
+        'MH1-M5'
+    """
+    gene = Parameter(gene, "gene").resolve_with_alias_and_return_value(
+        Parameter(gene_name, "gene_name")
+    )
+
+    Parameter(gene, "gene").throw_error_if_not_of_type(str)
+    Parameter(species, "species").throw_error_if_not_of_type(str)
+    Parameter(precision, "precision").throw_error_if_not_one_of(
+        "allele", "protein", "gene"
+    )
+    Parameter(on_fail, "on_fail").throw_error_if_not_one_of("reject", "keep")
+    Parameter(suppress_warnings, "suppress_warnings").throw_error_if_not_of_type(bool)
+
+    species = _utils.lowercase_and_remove_whitespace(species)
+
+    species_is_supported = species in STANDARDIZERS
+    if not species_is_supported:
+        if not suppress_warnings:
+            warnings.warn_unsupported_species(species, "MHC")
+        return gene
+
+    TcrStandardizerClass = STANDARDIZERS[species]
+    standardizer = TcrStandardizerClass(gene)
+
+    invalid_reason = standardizer.get_invalid_reason()
+    if invalid_reason is not None:
+        if not suppress_warnings:
+            warnings.warn_failure(
+                reason_for_failure=invalid_reason,
+                original_input=gene,
+                attempted_fix=standardizer.compile("allele"),
+                species=species,
+            )
+        if on_fail == "reject":
+            return None
+        return gene
+
+    return standardizer.compile(precision)
+
+
+def standardise(*args, **kwargs):
+    """
+    Alias for :py:func:`tidytcells.mhc.standardize`.
+    """
+    return standardize(*args, **kwargs)
